rustsbi: guest instance; u32 function and module id;

+++ src/ecall/ipi.rs
@@ -2,10 +2,10 @@ use super::SbiRet;
 use crate::hart_mask::HartMask;
 use crate::ipi::{max_hart_id, send_ipi_many};
 
-const FUNCTION_IPI_SEND_IPI: usize = 0x0;
+const FUNCTION_IPI_SEND_IPI: u32 = 0x0;
 
 #[inline]
-pub fn handle_ecall_ipi(function: usize, param0: usize, param1: usize) -> SbiRet {
+pub fn handle_ecall_ipi(function: u32, param0: usize, param1: usize) -> SbiRet {
     match function {
         FUNCTION_IPI_SEND_IPI => send_ipi(param0, param1),
         _ => SbiRet::not_supported(),
